suggestions, remove weird logic from _check_window
