+++ bach/tests/functional/bach/test_df_groupby.py
@@ -7,8 +7,7 @@ import numpy
 import pytest
 
 from bach import Series, SeriesAbstractNumeric
-from bach.partitioning import GroupingList, GroupingSet, Rollup, Cube
-from bach.utils import WindowFunction
+from bach.partitioning import GroupingList, GroupingSet, Rollup, Cube, WindowFunction
 from sql_models.util import is_postgres
 from tests.functional.bach.test_data_and_utils import assert_equals_data, get_df_with_test_data
